Factor the batch table hierarchy out of the old batch table
